external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/extensions/oes-texture-float-expected.txt

This test verifies the functionality of the OES_texture_float extension, if it is available.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS WebGL context exists
PASS getError was expected value: NO_ERROR : texture parameter setup should succeed
PASS getError was expected value: INVALID_ENUM : floating-point texture allocation must be disallowed if OES_texture_float isn't enabled
PASS No OES_texture_float support -- this is legal

PASS successfullyParsed is true

TEST COMPLETE
